Complete top_gems task
There is a program that fetch data about gems that lives on github.com
When user types on console "ruby top_gem.rb" the program takes names of
the gems from "gems.yml" file and store two arrays. One contains gems names
(the program going to take that array when it will build a header for console
output spreadshit) and second contains a uri github list of that gems,
received via "rubygems.org" api (all these manipulations happens in GemsFetch
class). Then the program goes to the github, parses two pages where desired data
lives and store it on two hashes (These happens in Extractor module). After that
program unites them to one hash that contains all data. Then it takes the hash and
array of gems names and store data in one array of arrays to convenient way for
console output spreadshit body. Also it build an array of hash keys for the
spreadshit header (all these happens in DataBuilder class). Finally the
program prints the spreadshit on the user monitor.

In addition user can type a bunch of console commands for top gems sorting and for
selecting the output the gems by name (or over bit of gems names). Also he
could take his own yaml file with a gems list.

In this case program will sort data by user demands. It will calculate a score
that consists of the sum of all gems parameters (stars, watchs etc.) or/and
select the data by matching over the bit of gems names.