Completed 2_task (telegram-bot @workflow_checker_bot)
When a user start this bot he would asked for his camp number
for registration. If the user would input incorrect number or
number that already taken by another user he would be alerted
by bot and would be offered to input the correct one.
After successful registration the user would be offered to make
/checkin command. Within this stage user would load his selfie
and geolocation coordinates. If his selfie wouldn't be an image
or his geolocation would be far from camp location he would be
politely told by the bot about this occurance and would be asked
for correct actions. If these actions happened the selfie and
geolocations would be nested to file with specific path with
timestamp mark.
The same result would be if the user would made a /checkout
command.
It is worth saying the user couldn't do /checkin command while
/checkout command is taking place and vice versa.